+++ oci/database_management_managed_databases_change_database_parameter_test.go
@@ -21,15 +21,20 @@ var (
 	}
 
 	managedDatabasesChangeDatabaseParameterCredentialsRepresentation = map[string]interface{}{
-		"password":  Representation{repType: Required, create: `system`},
-		"role":      Representation{repType: Required, create: `NORMAL`},
-		"user_name": Representation{repType: Required, create: `system`},
+		"password":  Representation{repType: Optional, create: `system`},
+		"role":      Representation{repType: Optional, create: `NORMAL`},
+		"secret_id": Representation{repType: Optional, create: `${oci_vault_secret.test_secret.id}`},
+		"user_name": Representation{repType: Optional, create: `system`},
 	}
 	managedDatabasesChangeDatabaseParameterParametersRepresentation = map[string]interface{}{
 		"name":           Representation{repType: Required, create: `open_cursors`},
 		"value":          Representation{repType: Required, create: `305`},
 		"update_comment": Representation{repType: Required, create: `Terraform provision of open cursors`},
 	}
+
+	ManagedDatabasesChangeDatabaseParameterResourceDependencies = generateDataSourceFromRepresentationMap("oci_database_management_managed_databases", "test_managed_databases", Required, Create, managedDatabaseDataSourceRepresentation) +
+		generateResourceFromRepresentationMap("oci_kms_vault", "test_vault", Required, Create, vaultRepresentation) +
+		generateDataSourceFromRepresentationMap("oci_vault_secrets", "test_secrets", Required, Create, secretDataSourceRepresentation)
 )
 
 // issue-routing-tag: database_management/default