+++ oci/database_management_managed_databases_reset_database_parameter_resource.go
@@ -48,6 +48,12 @@ func DatabaseManagementManagedDatabasesResetDatabaseParameterResource() *schema.
 							Computed: true,
 							ForceNew: true,
 						},
+						"secret_id": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Computed: true,
+							ForceNew: true,
+						},
 						"user_name": {
 							Type:     schema.TypeString,
 							Optional: true,
@@ -175,6 +181,11 @@ func (s *DatabaseManagementManagedDatabasesResetDatabaseParameterResourceCrud) m
 		result.Role = oci_database_management.DatabaseCredentialsRoleEnum(role.(string))
 	}
 
+	if secretId, ok := s.D.GetOkExists(fmt.Sprintf(fieldKeyFormat, "secret_id")); ok {
+		tmp := secretId.(string)
+		result.SecretId = &tmp
+	}
+
 	if userName, ok := s.D.GetOkExists(fmt.Sprintf(fieldKeyFormat, "user_name")); ok {
 		tmp := userName.(string)
 		result.UserName = &tmp
@@ -191,6 +202,10 @@ func DatabaseCredentialsToMap(obj *oci_database_management.DatabaseCredentials) 
 	}
 
 	result["role"] = string(obj.Role)
+
+	if obj.SecretId != nil {
+		result["secret_id"] = string(*obj.SecretId)
+	}
 
 	if obj.UserName != nil {
 		result["user_name"] = string(*obj.UserName)