+++ oci/database_management_managed_databases_reset_database_parameter_test.go
@@ -21,10 +21,15 @@ var (
 	}
 
 	managedDatabasesResetDatabaseParameterCredentialsRepresentation = map[string]interface{}{
-		"password":  Representation{repType: Required, create: "system"},
-		"role":      Representation{repType: Required, create: `NORMAL`},
-		"user_name": Representation{repType: Required, create: "system"},
+		"password":  Representation{repType: Optional, create: `system`},
+		"role":      Representation{repType: Optional, create: `NORMAL`},
+		"secret_id": Representation{repType: Optional, create: `${oci_vault_secret.test_secret.id}`},
+		"user_name": Representation{repType: Optional, create: `system`},
 	}
+
+	ManagedDatabasesResetDatabaseParameterResourceDependencies = generateDataSourceFromRepresentationMap("oci_database_management_managed_databases", "test_managed_databases", Required, Create, managedDatabaseDataSourceRepresentation) +
+		generateResourceFromRepresentationMap("oci_kms_vault", "test_vault", Required, Create, vaultRepresentation) +
+		generateDataSourceFromRepresentationMap("oci_vault_secrets", "test_secrets", Required, Create, secretDataSourceRepresentation)
 )
 
 // issue-routing-tag: database_management/default