+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support change compartment using lifecycle state
 - Support retry for Eventually consistency failure
 - Support for Count API added to `os_management` resource
+- Support for cloud databases in Database Management
 
 ## 4.42.0 (September 01, 2021)